Reorganizing components into proper subfolders
