+++ src/components/Applications/Applications.module.css
