+++ src/components/Applications/SingleApplication.module.css
