+++ src/components/Header/Header.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import LogoSvg from "./logo.svg";
+import LogoSvg from "../../logo.svg";
 import styles from "./Header.module.css";
 
 const Header = () => {
